Create overlays on-demand when focus lands on untracked window
tarmac multiplexes workspaces by hiding non-current windows; only
windows on the active workspace get enumerated by discover_windows
at startup. When the user switched to a workspace whose windows
weren't tracked, focus would move to a wid with no overlay and the
border would disappear (active-only mode hides the previous overlay
but has nothing to unhide for the new focus).

Trace from /tmp/ers-debug.log:
  [focus] 130 -> 139 (tracked targets: [796, 130, 681, 795, 778])
  [hide] target=130
  (no unhide — 139 not in tracked set, border vanishes)

Fix: in update_focus, if the new front isn't tracked, add_fresh it
on the spot. Same treatment for Event::Unhide so workspace-switch
unhide notifications can pick up newly-visible windows.

Also: CFRunLoopTimerCreate flags arg was u32, should be u64
(CFOptionFlags = unsigned long on 64-bit Darwin) — corrupted ABI for
subsequent args and the timer was firing unreliably as a result.

+++ src/skylight.rs
@@ -481,7 +481,7 @@ unsafe extern "C" {
         allocator: CFAllocatorRef,
         fire_date: f64,
         interval: f64,
-        flags: u32,
+        flags: u64,
         order: i64,
         callout: extern "C" fn(*mut c_void, *mut c_void),
         context: *mut CFRunLoopTimerContext,
